+++ src/main/java/racingcar/controller/RacingCarController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import racingcar.RaceDto;
+import racingcar.dto.RaceDto;
 import racingcar.dto.CarPositionDto;
 import racingcar.dto.CarResponse;
 import racingcar.dto.GameRequest;
